split installer steps
